pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "pymongo>=4.16.0",
    "python-jose>=3.5.0",
    "pytz>=2026.1.post1",
    "uuid>=1.30",
]
